@@ -4,15 +4,15 @@
   "description": "API boilerplate for datastore",
   "author": "DNA Micro",
   "license": "MIT",
-  "main": "./dist/main.js",
+  "main": "./src/main.ts",
   "scripts": {
     "prebuild": "rimraf dist",
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
-    "start:dev": "NODE_ENV='development' nest start --watch",
-    "start:debug": "NODE_ENV='development' DEBUG=true nest start --debug --watch",
-    "start:prod": "NODE_ENV='production' node dist/main",
+    "start:dev": "bun run build && bun --watch run ./dist/main.js",
+    "start:debug": "bun run build && DEBUG=true bun --watch --debug run ./dist/main.js",
+    "start:prod": "NODE_ENV='production' bun dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "start:test:prod": "bun run test && bun run build && bun run start:prod",
@@ -29,27 +29,25 @@
     "url": ""
   },
   "dependencies": {
-    "uuid": "^10.0.0",
+    "@dna-platform/common": "^1.0.14",
+    "@elastic/elasticsearch": "^8.10.0",
+    "@nestjs/axios": "^3.0.2",
+    "@nestjs/common": "^10.4.6",
+    "@nestjs/config": "^3.2.2",
+    "@nestjs/core": "^10.4.6",
+    "@nestjs/platform-express": "^10.4.6",
+    "@nestjs/serve-static": "^4.0.2",
+    "axios": "^1.5.0",
     "bluebird": "^3.7.2",
+    "classnames": "^2.3.2",
     "cookie-parser": "^1.4.7",
     "dotenv": "^16.4.5",
     "drizzle-orm": "^0.35.3",
     "drizzle-zod": "^0.5.1",
+    "express": "^4.21.1",
     "jsonwebtoken": "^9.0.2",
-    "murmurhash": "^2.0.1",
-    "ulid": "^2.3.0",
-    "zod": "^3.23.8",
-    "@dna-platform/common": "^1.0.0",
-    "@elastic/elasticsearch": "^8.10.0",
-    "@nestjs/axios": "^3.0.2",
-    "@nestjs/common": "^10.3.9",
-    "@nestjs/config": "^3.2.2",
-    "@nestjs/core": "^10.3.9",
-    "@nestjs/platform-express": "^10.3.9",
-    "@nestjs/serve-static": "^4.0.2",
-    "axios": "^1.5.0",
-    "classnames": "^2.3.2",
     "lowdb": "1.0.0",
+    "murmurhash": "^2.0.1",
     "nest-winston": "^1.9.4",
     "node-mocks-http": "^1.14.1",
     "object-path": "^0.11.8",
@@ -57,10 +55,14 @@
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
     "split2": "^4.2.0",
+    "sqlite3": "^5.1.7",
     "typed-html": "^3.0.1",
     "uglify-js": "^3.17.4",
+    "ulid": "^2.3.0",
+    "uuid": "^10.0.0",
     "winston": "^3.10.0",
-    "xstate": "^5.16.0"
+    "xstate": "^5.16.0",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@nestjs/cli": "^10.3.2",
@@ -68,13 +70,12 @@
     "@nestjs/testing": "^10.3.9",
     "@types/bun": "^1.1.12",
     "@types/cookie-parser": "^1.4.7",
-    "@types/express": "^4.17.13",
+    "@types/express": "^5.0.0",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.14.9",
     "@types/supertest": "^6.0.2",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
-    "bun-types": "^1.1.33",
     "drizzle-kit": "^0.26.2",
     "eslint": "^8.0.1",
     "eslint-config-prettier": "^8.3.0",
@@ -87,7 +88,7 @@
     "ts-loader": "^9.2.3",
     "ts-node": "^10.0.0",
     "tsconfig-paths": "4.0.0",
-    "typescript": "^5.5.4"
+    "typescript": "^5.6.3"
   },
   "jest": {
     "moduleFileExtensions": [
