@@ -8,5 +8,5 @@ import { defineConfig } from 'drizzle-kit';
 */
 export default defineConfig({
   dialect: 'sqlite',
-  schema: './src/schema/index.ts',
+  schema: './src/sync/schema/index.ts',
 });
