@@ -5,7 +5,6 @@ import * as schema from '../../schema';
 import { migrate } from 'drizzle-orm/bun-sqlite/migrator';
 const { DB_FILE_DIR = 'sql', DB_FILE_SQLITE = 'sqlite.db' } = process.env;
 
-
 @Injectable()
 export class DrizzleService {
   private logger = new Logger(DrizzleService.name);
@@ -15,7 +14,7 @@ export class DrizzleService {
   constructor() {}
   async onModuleInit() {
     this.logger.log('Loading SQL File' + DB_FILE_DIR + '/' + DB_FILE_SQLITE);
-    await migrate(this.db, { migrationsFolder: './drizzle' });
+    migrate(this.db, { migrationsFolder: './drizzle' });
   }
 
   getClient() {
