@@ -6,14 +6,14 @@ export * from './crdt/queue';
 export * from './crdt/sync_endpoints';
 
 // export all from ./application folder
-export * from './application/contact_emails.schema';
-export * from './application/contact_phone_numbers.schema';
-export * from './application/contacts.schema';
+export * from './application/contact_emails';
+export * from './application/contact_phone_numbers';
+export * from './application/contacts';
 export * from './application/organization_contact_accounts.schema';
 export * from './application/organization_contacts.schema';
 export * from './application/organization_domains.schema';
 export * from './application/organization_files.schema';
 export * from './application/organizations.schema';
-export * from './application/config_applications.schema';
+export * from './application/config_applications';
 
 export * from './config/config_sync.schema';
